chore:  more minimal
